requirements.txt
fastapi[all]
uvicorn[standard]
markdown
gunicorn
leapcell
python-dotenv